[test] Move all pmod tests to a spearate folder.
Move all pmod test cases under //sw/device/tests/pmod to make it more
obvious that these tests require additional test harness support.

A README describing the pmod test harness can be added later.
