+++ sw/device/tests/pmod/spi_host_gigadevice1Gb_flash_test.c
@@ -3,11 +3,11 @@
 // SPDX-License-Identifier: Apache-2.0
 #include <assert.h>
 
-#include "spi_host_flash_test_impl.h"
 #include "sw/device/lib/arch/device.h"
 #include "sw/device/lib/base/macros.h"
 #include "sw/device/lib/base/memory.h"
 #include "sw/device/lib/base/mmio.h"
+#include "sw/device/lib/dif/dif_pinmux.h"
 #include "sw/device/lib/dif/dif_spi_host.h"
 #include "sw/device/lib/runtime/hart.h"
 #include "sw/device/lib/runtime/log.h"
@@ -17,6 +17,7 @@
 #include "sw/device/lib/testing/spi_host_testutils.h"
 #include "sw/device/lib/testing/test_framework/check.h"
 #include "sw/device/lib/testing/test_framework/ottf_main.h"
+#include "sw/device/tests/spi_host_flash_test_impl.h"
 
 #include "hw/top_earlgrey/sw/autogen/top_earlgrey.h"
 
@@ -31,7 +32,7 @@ static void init_test(dif_spi_host_t *spi_host) {
       mmio_region_from_addr(TOP_EARLGREY_PINMUX_AON_BASE_ADDR);
   CHECK_DIF_OK(dif_pinmux_init(base_addr, &pinmux));
   CHECK_STATUS_OK(
-      spi_host1_pinmux_connect_to_bob(&pinmux, kTopEarlgreyPinmuxMioOutIob7));
+      spi_host1_pinmux_connect_to_bob(&pinmux, kTopEarlgreyPinmuxMioOutIoc6));
 
   base_addr = mmio_region_from_addr(TOP_EARLGREY_SPI_HOST1_BASE_ADDR);
   CHECK_DIF_OK(dif_spi_host_init(base_addr, spi_host));
@@ -54,10 +55,18 @@ bool test_main(void) {
   dif_spi_host_t spi_host;
 
   init_test(&spi_host);
-  enum MacronixVendorSpecific {
-    kDeviceId = 0x1B20,
-    kManufactureId = 0xC2,
-    kPageQuadProgramOpcode = 0x38,
+
+  /**
+   * This flash device does not support dual mode read, so we don't test it.
+   * It also does not require a command to enable the quad mode. Unless we want
+   * to use the QPI (Quad Peripheral Interface) mode. QPI differs from Quad SPI
+   * mode on the opcode section that use four lanes as opposed to one lane in
+   * the Quad SPI.
+   */
+  enum GigadeviceVendorSpecific {
+    kDeviceId = 0x1b47,
+    kManufactureId = 0xC8,
+    kPageQuadProgramOpcode = 0xC2,
   };
 
   status_t result = OK_STATUS();
@@ -65,15 +74,14 @@ bool test_main(void) {
   EXECUTE_TEST(result, test_read_sfdp, &spi_host);
   EXECUTE_TEST(result, test_sector_erase, &spi_host);
   EXECUTE_TEST(result, test_read_jedec, &spi_host, kDeviceId, kManufactureId);
-  EXECUTE_TEST(result, test_enable_quad_mode, &spi_host);
   EXECUTE_TEST(result, test_page_program, &spi_host);
   if (is_4_bytes_address_mode_supported()) {
     EXECUTE_TEST(result, test_4bytes_address, &spi_host);
   }
   EXECUTE_TEST(result, test_fast_read, &spi_host);
-  EXECUTE_TEST(result, test_dual_read, &spi_host);
   EXECUTE_TEST(result, test_quad_read, &spi_host);
-  // The Macronix flash `4PP` opcode operates in 1-4-4 mode.
+
+  // The Gigadevice flash `4PP` opcode operates in 1-4-4 mode.
   EXECUTE_TEST(result, test_page_program_quad, &spi_host,
                kPageQuadProgramOpcode, kTransactionWidthMode144);
   EXECUTE_TEST(result, test_erase_32k_block, &spi_host);
