+++ sw/device/tests/pmod/spi_host_macronix_flash_test.c
@@ -3,12 +3,10 @@
 // SPDX-License-Identifier: Apache-2.0
 #include <assert.h>
 
-#include "spi_host_flash_test_impl.h"
 #include "sw/device/lib/arch/device.h"
 #include "sw/device/lib/base/macros.h"
 #include "sw/device/lib/base/memory.h"
 #include "sw/device/lib/base/mmio.h"
-#include "sw/device/lib/dif/dif_pinmux.h"
 #include "sw/device/lib/dif/dif_spi_host.h"
 #include "sw/device/lib/runtime/hart.h"
 #include "sw/device/lib/runtime/log.h"
@@ -18,6 +16,7 @@
 #include "sw/device/lib/testing/spi_host_testutils.h"
 #include "sw/device/lib/testing/test_framework/check.h"
 #include "sw/device/lib/testing/test_framework/ottf_main.h"
+#include "sw/device/tests/spi_host_flash_test_impl.h"
 
 #include "hw/top_earlgrey/sw/autogen/top_earlgrey.h"
 
@@ -32,7 +31,7 @@ static void init_test(dif_spi_host_t *spi_host) {
       mmio_region_from_addr(TOP_EARLGREY_PINMUX_AON_BASE_ADDR);
   CHECK_DIF_OK(dif_pinmux_init(base_addr, &pinmux));
   CHECK_STATUS_OK(
-      spi_host1_pinmux_connect_to_bob(&pinmux, kTopEarlgreyPinmuxMioOutIoc10));
+      spi_host1_pinmux_connect_to_bob(&pinmux, kTopEarlgreyPinmuxMioOutIob7));
 
   base_addr = mmio_region_from_addr(TOP_EARLGREY_SPI_HOST1_BASE_ADDR);
   CHECK_DIF_OK(dif_spi_host_init(base_addr, spi_host));
@@ -55,16 +54,10 @@ bool test_main(void) {
   dif_spi_host_t spi_host;
 
   init_test(&spi_host);
-
-  /**
-   * This flash has bug where it does not report the busy bit on the status
-   * register correctly for the quad page program. So we would have to use a non
-   * standard method (flag status register). So we won't test quad test program
-   * here.
-   */
-  enum MicronVendorSpecific {
-    kDeviceId = 0x20ba,
-    kManufactureId = 0x20,
+  enum MacronixVendorSpecific {
+    kDeviceId = 0x1B20,
+    kManufactureId = 0xC2,
+    kPageQuadProgramOpcode = 0x38,
   };
 
   status_t result = OK_STATUS();
@@ -80,6 +73,9 @@ bool test_main(void) {
   EXECUTE_TEST(result, test_fast_read, &spi_host);
   EXECUTE_TEST(result, test_dual_read, &spi_host);
   EXECUTE_TEST(result, test_quad_read, &spi_host);
+  // The Macronix flash `4PP` opcode operates in 1-4-4 mode.
+  EXECUTE_TEST(result, test_page_program_quad, &spi_host,
+               kPageQuadProgramOpcode, kTransactionWidthMode144);
   EXECUTE_TEST(result, test_erase_32k_block, &spi_host);
   EXECUTE_TEST(result, test_erase_64k_block, &spi_host);
